+++ src/main/java/ru/rsreu/expertsandteams/service/SessionService.java
@@ -5,5 +5,5 @@ import ru.rsreu.expertsandteams.model.entity.Session;
 import java.util.Optional;
 
 public interface SessionService {
-    Optional<Session> getSession(int userId);
+    Optional<Session> getSession(Long userId);
 }
